Model and completion for property expressions
In many MicroProfile implementations, `${other.property}` can be
used to refer to the value of an already defined property when defining
the value of another property.

This PR modifies the property file model to allow representing these
property expressions.
It also adds completion for defined properties when completion is
opened after typing `${` as a part of a property value.

This PR was migrated from
[quarkus-ls](https://github.com/redhat-developer/quarkus-ls/pull/340).

Signed-off-by: David Thompson <davthomp@redhat.com>