Extension point for contributing configuration sources
lsp4mp extensions can now contribute implementations of IConfigSourceProvider,
which is a class that generates a list of IConfigSource given a project.
All the IConfigSource generated by all the IConfigSourceProvider
will be used for language features that use a property's value
(eg. hover).

This means that if an implementation of MicroProfile provides
a particular configuration file that is not a part of the base MicroProfile config,
an extension to lsp4mp can be implemented which introduces support for
the configuration file, without having to modify lsp4mp directly.

As example of how this can be used, this opens the door to moving
the implementation and inclusion of the Quarkus
`application.properties` and `application.yml` configuration sources out of
lsp4mp, and instead providing them through a seperate project.

Closes redhat-developer/quarkus-ls#373

Signed-off-by: David Thompson <davthomp@redhat.com>