Support CodeAction resolve for Java code actions
Resolve workspace edits only when the code action is applied
for all code actions for Java files that modify source files.
This should increase the performance of editing Java files with lsp4mp
active, since the edits of all the code actions don't need to be
computed.

I also adjusted the title of the "Add @Operation annotation" code action
to include the class that the code action will apply to.

Closes #171

Signed-off-by: David Thompson <davthomp@redhat.com>