Revert "Use JavadocContentAccess2"
This reverts commit b7d2f9520a8b2eb545ceb9e3952d2823dc322fba.

The change I made included code to collect Javadoc for all the
properties. However, since all the Javadoc was being collected when the
language server started up, it would cause large projects to take much
longer to open. I will handle collecting the Javadoc for properties
properly in a future PR.

In the meantime, switching from `JavadocContentAccess` to
`JavadocContentAccess2` is not that important, since the method where
it's used isn't used much in lsp4mp.