Show/hide Java snippets based on cursor location
Snippet completion sends a request to the java language server to
determine additional context based on the cursor location and content of
the java file, then filters the snippets based on this context.

The context is fetched as a part of the existing
`microprofile/java/completion` request.

This means that:
- snippets that generate class declarations (like `rest_class`,
  `mpreadiness`) only appear in empty files
- snippets that generate an annotation only appear when the cursor is
  in a class and before a declaration that the given annotation can be
  applied to.

This PR also removes code related to enabling/disabling the feature of
getting the completion results directly from the JDT component.

Closes #108

Signed-off-by: David Thompson <davthomp@redhat.com>