Extension point to customize JAX-RS features
Allows you to customize the JAX-RS/Jakarta RESTful features using a new
extension point.
lsp4mp uses any implementations of `IJaxRsInfoParticipant` that are
contributed through external extensions onto lsp4mp in order to
collect a list of all classes that contain JAX-RS methods.
When collecting JAX-RS methods for a given class file, lsp4mp uses
the first `IJaxRsInfoParticipant` that reports it is able to collect
JAX-RS methods for a given class in order to do so.
If no `IJaxRsInfoParticipant` can resolve JAX-RS methods for a class,
it uses the default semantics for resolving JAX-RS methods.

This way, if you are working within a framework that uses custom logic
for determining which methods are REST endpoints,
or custom logic for what the corresponding URLs are for those endpoints,
lsp4mp can support that use case through additional contributed logic.

Signed-off-by: David Thompson <davthomp@redhat.com>