Add an `id` to track the type of each `CodeAction`
Add a new `id` field in the CodeAction data to identify
the type of each CodeAction, without having to rely
on the display text of the code action.

This is very helpful for telemetry, since it enables
collecting info on which CodeActions are the most used,
while avoid pitfalls that could result in personal information being
leaked.

Signed-off-by: David Thompson <davthomp@redhat.com>